+++ packages/gatsby-transformer-cloudinary/get-image-objects/get-shared-image-data.js
@@ -52,15 +52,16 @@ exports.getAspectRatio = (transformations, originalAspectRatio) => {
 };
 
 exports.getBase64 = async ({
+  base64Transformations,
+  base64Width,
+  chained,
+  cloudName,
   defaultBase64,
   ignoreDefaultBase64,
   public_id,
-  version,
-  cloudName,
-  base64Transformations,
+  reporter,
   transformations,
-  base64Width,
-  chained,
+  version,
 }) => {
   if (defaultBase64) {
     if (!ignoreDefaultBase64 || getPluginOptions().alwaysUseDefaultBase64) {
@@ -76,17 +77,59 @@ exports.getBase64 = async ({
     cloudName,
     chained,
   });
-  const base64 = await fetchBase64(base64URL);
+  const base64 = await fetchBase64(base64URL, reporter);
 
   return base64;
 };
 
-async function fetchBase64(url) {
+async function fetchBase64(url, reporter) {
   if (!base64Cache[url]) {
+    logBase64Retrieval(url, reporter);
     const result = await axios.get(url, { responseType: 'arraybuffer' });
     const data = Buffer.from(result.data).toString('base64');
     base64Cache[url] = `data:image/jpeg;base64,${data}`;
   }
 
   return base64Cache[url];
+}
+
+let fetchedBase64ImageCount = 0;
+
+function logBase64Retrieval(url, reporter) {
+  fetchedBase64ImageCount += 1;
+  if (typeof reporter.info === 'function') {
+    reporter.info(
+      `[gatsby-transformer-cloudinary] Fetching base64 image ` +
+        `#${fetchedBase64ImageCount} from Cloudinary: ${url}`,
+    );
+    if (fetchedBase64ImageCount == 100) {
+      reporter.info(
+        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~',
+      );
+      reporter.info(
+        'It looks like your project has a lot of images. To improve your build times, ',
+      );
+      reporter.info(
+        'you should consider (1) using images that are already on Cloudinary, (2) ',
+      );
+      reporter.info(
+        'precomputing the base64 images and providing them to this plugin as the ',
+      );
+      reporter.info(
+        '`defaultBase64` property, and (3) setting the plugin option ',
+      );
+      reporter.info(
+        '`alwaysUseDefaultBase64` to true. Doing so will reduce the number of base64 ',
+      );
+      reporter.info(
+        'images that need to be fetched from Cloudinary and speed up your Gatsby "query" ',
+      );
+      reporter.info(
+        'build steps. (See the section "Use images already on Cloudinary" in the README.)',
+      );
+      reporter.info(
+        '[gatsby-transformer-cloudinary] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~',
+      );
+    }
+  }
 }