+++ packages/gatsby-transformer-cloudinary/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version Next
+
+Additions:
+
+- Added logging for each time we have to fetch a base64 image from Cloudinary to explain long query steps in the Gatsby build process.
+
 # Version 2.1.0
 
 Additions: